# docs/swagger.yaml
definitions:
  api.cancel_order_request:
    properties:
      order_id:
        type: string
    type: object
  api.new_order_request:
    properties:
      amount:
        example: "100.00"
        type: string
      order_type:
        example: limit/market
        type: string
      price:
        example: "1.00"
        type: string
      quantity:
        example: "12"
        type: string
      side:
        example: sell/buy
        type: string
      symbol:
        example: ethusd
        type: string
    required:
    - order_type
    - side
    - symbol
    type: object
  common.Response:
    properties:
      data: {}
      ok:
        type: integer
      reason:
        type: string
    type: object
  models.Kline:
    properties:
      amount:
        description: 成交额
        type: string
      close:
        description: 收盘价(当前K线未结束的即为最新价)
        type: string
      close_at:
        description: 收盘时间
        type: string
      high:
        description: 最高价
        type: string
      low:
        description: 最低价
        type: string
      open:
        description: 开盘价
        type: string
      open_at:
        description: 开盘时间
        type: string
      trade_count:
        description: 成交笔数
        type: integer
      volume:
        description: 成交量
        type: string
    type: object
  orders.TradeRecord:
    properties:
      amount:
        type: string
      ask:
        type: string
      bid:
        type: string
      create_time:
        type: string
      price:
        type: string
      quantity:
        type: string
      trade_by:
        type: integer
      trade_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/depth:
    get:
      parameters:
      - description: 'eg: ethusd'
        in: query
        name: symbol
        required: true
        type: string
      - description: 默认100，最大5000
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 深度信息
  /api/v1/market/klines:
    get:
      consumes:
      - application/json
      description: 行情k线数据接口
      parameters:
      - description: 'eg: ethusd'
        in: query
        name: symbol
        required: true
        type: string
      - description: m1,m3,m5...
        in: query
        name: period
        required: true
        type: string
      - description: 'default: 10'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kline'
      summary: 行情k线数据
      tags:
      - 行情接口
  /api/v1/order:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
  /api/v1/order/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
  /api/v1/order/cancel:
    post:
      consumes:
      - application/json
      description: 取消还未完成的订单
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/api.cancel_order_request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: 取消委托
      tags:
      - 订单相关
  /api/v1/order/new:
    post:
      consumes:
      - application/json
      description: |-
        新订单，支持限价单、市价单
        不同订单类型的参数要求：
        限价单: {"symbol": "ethusd", "order_type": "limit", "side": "sell", "price": "1.00", "quantity": "100"}
        市价-按数量: {"symbol": "ethusd", "order_type": "market", "side": "sell", "quantity": "100"}
        市价-按金额: {"symbol": "ethusd", "order_type": "market", "side": "sell", "amount": "1000.00"}
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/api.new_order_request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: 委托订单
      tags:
      - 订单相关
  /api/v1/order/open:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
  /api/v1/trade/record:
    get:
      parameters:
      - description: 'eg: ethusd'
        in: query
        name: symbol
        required: true
        type: string
      - description: 默认10，最大100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/orders.TradeRecord'
            type: array
      summary: 成交记录
swagger: "2.0"